requirements.txt
torch
numpy
opencv-python
imutils
Flask
mysql-connector-python
huggingface_hub
ultralytics
Pillow
paddlepaddle
paddleocr